Almost Complete

+++ XslxExporter/XslxExporter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
 using Common;
@@ -21,6 +22,11 @@ namespace XslxExporter
         public string Shortcut => "XLSX";
         public string Filter => "*.xlsx|*.xlsx";
 
+        public async Task ExportAsync(IEnumerable<Profile> profiles, ExportSettings settings)
+        {
+            await Task.Run(() => Export(profiles, settings)).ConfigureAwait(false);
+        }
+
         private void MakeHeaders(ExcelWorksheet ws, ExportSettings s)
         {
             const int row = 1;
@@ -30,11 +36,13 @@ namespace XslxExporter
                 ws.Cells[row, col].Value = "РШФ Id";
                 col++;
             }
+
             if (s.Fide)
             {
                 ws.Cells[row, col].Value = "Fide Id";
                 col++;
             }
+
             if (s.RuName && s.EngName)
             {
                 ws.Cells[row, col].Value = "Имя (Ру)";
@@ -47,121 +55,128 @@ namespace XslxExporter
                 ws.Cells[row, col].Value = "Имя";
                 col++;
             }
+
             if (s.Birth)
             {
                 ws.Cells[row, col].Value = "Год рожд.";
                 col++;
             }
+
             if (s.RcfRat[0])
             {
                 ws.Cells[row, col].Value = "РШФ Классика";
                 col++;
             }
+
             if (s.RcfRat[1])
             {
                 ws.Cells[row, col].Value = "РШФ Рапид";
                 col++;
             }
+
             if (s.RcfRat[2])
             {
                 ws.Cells[row, col].Value = "РШФ Блиц";
                 col++;
             }
+
             if (s.FideRat[0])
             {
                 ws.Cells[row, col].Value = "Fide Классика";
                 col++;
             }
+
             if (s.FideRat[1])
             {
                 ws.Cells[row, col].Value = "Fide Рапид";
                 col++;
             }
+
             if (s.FideRat[2])
             {
                 ws.Cells[row, col].Value = "Fide Блиц";
                 col++;
             }
-            if (s.Groups)
-            {
-                ws.Cells[row, col].Value = "Группа";
-            }
+
+            if (s.Groups) ws.Cells[row, col].Value = "Группа";
         }
 
         private void WriteRow(ExcelWorksheet ws, Profile pr, ExportSettings s, int row)
         {
-                var col = 1;
-                if (s.Rcf)
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.RcfId ?? 0;
-                    col++;
-                }
-                if (s.Fide)
-                {
-                    ws.Cells[row, col].Value = pr.FideProfile?.FideId ?? 0;
-                    col++;
-                }
-                if (s.RuName && s.EngName)
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.Name ?? "";
-                    col++;
-                    ws.Cells[row, col].Value = pr.FideProfile?.Name ?? "";
-                    col++;
-                }
+            var col = 1;
+            if (s.Rcf)
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.RcfId ?? 0;
+                col++;
+            }
+
+            if (s.Fide)
+            {
+                ws.Cells[row, col].Value = pr.FideProfile?.FideId ?? 0;
+                col++;
+            }
+
+            if (s.RuName && s.EngName)
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.Name ?? "";
+                col++;
+                ws.Cells[row, col].Value = pr.FideProfile?.Name ?? "";
+                col++;
+            }
+            else
+            {
+                string name;
+                if (s.RuName)
+                    name = pr.RcfProfile?.Name ?? pr.FideProfile?.Name ?? "";
                 else
-                {
-                    string name;
-                    if (s.RuName)
-                    {
-                        name = pr.RcfProfile?.Name ?? pr.FideProfile?.Name ?? "";
-                    }
-                    else
-                    {
-                        name = pr.FideProfile?.Name ?? pr.RcfProfile?.Name ?? "";
-                    }
-                    ws.Cells[row, col].Value = name;
-                    col++;
-                }
-                if (s.Birth)
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.Birth ?? pr.FideProfile?.Birth ?? 0;
-                    col++;
-                }
-                if (s.RcfRat[0])
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.Std ?? 0;
-                    col++;
-                }
-                if (s.RcfRat[1])
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.Rpd ?? 0;
-                    col++;
-                }
-                if (s.RcfRat[2])
-                {
-                    ws.Cells[row, col].Value = pr.RcfProfile?.Blz ?? 0;
-                    col++;
-                }
-                if (s.FideRat[0])
-                {
-                    ws.Cells[row, col].Value = pr.FideProfile?.Std ?? 0;
-                    col++;
-                }
-                if (s.FideRat[1])
-                {
-                    ws.Cells[row, col].Value = pr.FideProfile?.Rpd ?? 0;
-                    col++;
-                }
-                if (s.FideRat[2])
-                {
-                    ws.Cells[row, col].Value = pr.FideProfile?.Blz ?? 0;
-                    col++;
-                }
-                if (s.Groups)
-                {
-                    ws.Cells[row, col].Value = pr.Group.Name;
-                }
-            
+                    name = pr.FideProfile?.Name ?? pr.RcfProfile?.Name ?? "";
+                ws.Cells[row, col].Value = name;
+                col++;
+            }
+
+            if (s.Birth)
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.Birth ?? pr.FideProfile?.Birth ?? 0;
+                col++;
+            }
+
+            if (s.RcfRat[0])
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.Std ?? 0;
+                col++;
+            }
+
+            if (s.RcfRat[1])
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.Rpd ?? 0;
+                col++;
+            }
+
+            if (s.RcfRat[2])
+            {
+                ws.Cells[row, col].Value = pr.RcfProfile?.Blz ?? 0;
+                col++;
+            }
+
+            if (s.FideRat[0])
+            {
+                ws.Cells[row, col].Value = pr.FideProfile?.Std ?? 0;
+                col++;
+            }
+
+            if (s.FideRat[1])
+            {
+                ws.Cells[row, col].Value = pr.FideProfile?.Rpd ?? 0;
+                col++;
+            }
+
+            if (s.FideRat[2])
+            {
+                ws.Cells[row, col].Value = pr.FideProfile?.Blz ?? 0;
+                col++;
+            }
+
+            if (s.Groups) ws.Cells[row, col].Value = pr.Group.Name;
         }
 
         private void Export(IEnumerable<Profile> profiles, ExportSettings settings)
@@ -179,31 +194,25 @@ namespace XslxExporter
             };
             if (dlg.ShowDialog() != true) return;
             var path = dlg.FileName;
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
+            if (File.Exists(path)) File.Delete(path);
             var xlsx = new ExcelPackage(new FileInfo(path));
             var ws = xlsx.Workbook.Worksheets.Add("RTG-" + DateTime.Today.Day + "-" + DateTime.Today.Month + "-" +
                                                   DateTime.Today.Year);
             var row = 2;
-            MakeHeaders(ws,settings);
+            MakeHeaders(ws, settings);
             foreach (var profile in profiles)
             {
-                WriteRow(ws,profile,settings,row);
+                WriteRow(ws, profile, settings, row);
                 row++;
             }
-            var range = ws.Cells[ws.Dimension.Start.Row, ws.Dimension.Start.Column, ws.Dimension.End.Row, ws.Dimension.End.Column];
+
+            var range = ws.Cells[ws.Dimension.Start.Row, ws.Dimension.Start.Column, ws.Dimension.End.Row,
+                ws.Dimension.End.Column];
             var table = ws.Tables.Add(range, "table1");
             range.AutoFitColumns();
             table.TableStyle = TableStyles.Light1;
             xlsx.Save();
-            System.Diagnostics.Process.Start(path);
-        }
-
-        public async Task ExportAsync(IEnumerable<Profile> profiles, ExportSettings settings)
-        {
-            await Task.Run(() => Export(profiles, settings)).ConfigureAwait(false);
+            Process.Start(path);
         }
     }
 }
